5: Move necessary project files out of src file for env issues and deploying setup
