+++ backend/index.ts
@@ -1,7 +1,12 @@
 import express from 'express';
+import * as dotenv from 'dotenv';
+dotenv.config();
 import routes from './routes/index.js';
 import usersRouter from './routes/users.routes.js';
 import cardsRouter from './routes/creditCards.routes.js';
+
+console.log('process', process.env.REDIS_URL);
+console.log('chris', process.env.DEV);
 
 const app = express();
 const port = 5000;
