+++ backend/package.json
@@ -7,15 +7,16 @@
   "type": "module",
   "dependencies": {
     "@types/express": "^4.17.13",
+    "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "nodemon": "^2.0.19",
     "redis-om": "^0.3.6",
     "ts-node": "^10.9.1"
   },
   "scripts": {
-    "dev": "cd src && nodemon index.ts",
+    "dev": "nodemon index.ts",
     "start": "tsc && node dist/app.js",
-	"redis": "redis-stack-server"
+    "redis": "redis-stack-server"
   },
   "devDependencies": {
     "typescript": "^4.7.4"
